backend/requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2019.3.9
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
daphne==2.2.5
dj-database-url==0.5.0
Django==2.1.7
django-extensions==2.1.6
django-heroku==0.3.1
djangorestframework==3.9.1
exponent-server-sdk==0.2.0
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyHamcrest==1.9.0
pytz==2018.9
requests==2.21.0
sentry-sdk==0.7.11
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.2
whitenoise==4.1.2
zope.interface==4.6.0
eventlet==0.24.1